% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-methods.R
\docType{methods}
\name{aggregateTree,TreeSE-method}
\alias{aggregateTree,TreeSE-method}
\title{Method to aggregate a TreeSE object}
\usage{
\S4method{aggregateTree}{TreeSE}(x, selectedLevel = 3,
  selectedNodes = NULL, aggFun = colSums, start = 1, end = NULL,
  by = "row", format = "TreeSE")
}
\arguments{
\item{x}{TreeSE object}

\item{selectedLevel}{level to select nodes from}

\item{selectedNodes}{used to set states on individual nodes to define a cut on the tree}

\item{aggFun}{aggregate function to use, by default colSums if by="row", rowSums if by="col"}

\item{start, end}{indices to filter nodes}

\item{by}{"row" to aggregate the TreeIndex on rowData, "col" to aggregate TreeIndex on colData}

\item{format}{return format can be one of "counts" or "TreeSE"}
}
\description{
Method to aggregate a TreeSE object
}
