% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeViz-methods.R
\name{aggregateTree,TreeViz-method}
\alias{aggregateTree,TreeViz-method}
\title{Method to aggregate a TreeViz object}
\usage{
\S4method{aggregateTree}{TreeViz}(
  x,
  selectedLevel = 3,
  selectedNodes = NULL,
  aggFun = colSums,
  start = 1,
  end = NULL,
  by = "row",
  format = "TreeViz"
)
}
\arguments{
\item{x}{TreeViz object}

\item{selectedLevel}{level to select nodes from}

\item{selectedNodes}{used to set states on individual nodes to define a cut on the tree}

\item{aggFun}{aggregate function to use, by default colSums if by="row", rowSums if by="col"}

\item{start, end}{indices to filter nodes}

\item{by}{"row" to aggregate the TreeIndex on rowData, "col" to aggregate TreeIndex on colData}

\item{format}{return format can be one of "counts" or "TreeViz"}
}
\description{
Method to aggregate a TreeViz object
}
\examples{
library(metagenomeSeq)
data(mouseData)
counts <- MRcounts(mouseData)
hierarchy <- fData(mouseData)
tree <- TreeIndex(hierarchy)
mbiome <- TreeViz(SimpleList(counts=counts), rowData=tree)
aggregateTree(mbiome)
}
