% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-methods.R
\name{aggregateTree}
\alias{aggregateTree}
\title{Generic method to aggregateTree}
\usage{
aggregateTree(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters}
}
\description{
Generic method to aggregateTree
}
