% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{calculate_tsne}
\alias{calculate_tsne}
\title{Calculates `tsne` Dimensionality Reduction on `TreeViz` object.}
\usage{
calculate_tsne(treeviz)
}
\arguments{
\item{treeviz}{TreeViz object}
}
\value{
`TreeViz` Object with added `TSNE` iinformation in reduced_dim slot of metadata
}
\description{
The result is added to reduced_dim slot in metadata
}
\examples{
\dontrun{
library(Seurat)
data(pbmc_small)
pbmc <- pbmc_small
treeviz<- createFromSeurat(pbmc, check_metadata = TRUE, reduced_dim = c("pca","tsne"))
treeviz<- calculate_tsne(treeviz)
metadata(treeviz)$reduced_dim$TSNE
}
}
