% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{change_assignment}
\alias{change_assignment}
\title{propagate cluster assignment changes from graph object to hierarchy}
\usage{
change_assignment(graph, cluster_obj)
}
\arguments{
\item{graph}{graph from clustree}

\item{cluster_obj}{cluster hierarchies at different resolutions}
}
\value{
hierarchy of cluster resolutions
}
\description{
propagate cluster assignment changes from graph object to hierarchy
}
