% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_alternate}
\alias{check_alternate}
\title{check if alternative edge exists}
\usage{
check_alternate(sub_df, all_df)
}
\arguments{
\item{sub_df}{dataframe}

\item{all_df}{dataframe}
}
\value{
Dataframe with assigned edges
}
\description{
check if alternative edge exists
}
