% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_cycle}
\alias{check_cycle}
\title{check if graph is cyclic}
\usage{
check_cycle(pruned_graph)
}
\arguments{
\item{pruned_graph}{graph from clustree}
}
\value{
graph object after removing cyclic edges
}
\description{
check if graph is cyclic
}
