% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_unique_parent}
\alias{check_unique_parent}
\title{check if every node has a single parent (tree vs graph)}
\usage{
check_unique_parent(clusterdata)
}
\arguments{
\item{clusterdata}{Dataframe containing cluster information at different resolutions}
}
\value{
Dataframe containing cluster information at different resolutions
}
\description{
check if every node has a single parent (tree vs graph)
}
