% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{checkRoot}
\alias{checkRoot}
\title{check if hierarchy has more than 1 root, if do create a root}
\usage{
checkRoot(cluster_df)

checkRoot(cluster_df)
}
\arguments{
\item{cluster_df}{Dataframe containing cluster information at different resolutions}
}
\value{
Dataframe containing cluster information at different resolutions

Dataframe containing cluster information at different resolutions
}
\description{
check if hierarchy has more than 1 root, if do create a root

check if hierarchy has more than 1 root, if do create a root
}
