% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCluster-class.R
\docType{class}
\name{ClusterHierarchy-class}
\alias{ClusterHierarchy-class}
\title{ClusterHierarchy class to manage treeviz cluster data}
\description{
ClusterHierarchy class to manage treeviz cluster data
}
