% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCluster-class.R
\name{ClusterHierarchy}
\alias{ClusterHierarchy}
\title{create a new ClusterHierarchy object}
\usage{
ClusterHierarchy(hierarchy, col_regex = NULL, columns = NULL)
}
\arguments{
\item{hierarchy}{hierarchy as a dataFrame}

\item{col_regex}{Regular Expression for choosing columns}

\item{columns}{Vector containing list of columns to choose from with ordering}
}
\value{
`ClusterHierarchy`` return an object of class ClusterHierarchy containing cluster information
}
\description{
create a new ClusterHierarchy object
}
