% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{collapse_tree}
\alias{collapse_tree}
\title{collapse tree if resolutions does not change cluster assignments}
\usage{
collapse_tree(original_graph)
}
\arguments{
\item{original_graph}{Dataframe containing cluster information at different resolutions}
}
\value{
Dataframe containing cluster information at different resolutions
}
\description{
collapse tree if resolutions does not change cluster assignments
}
