% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{createFromSCE}
\alias{createFromSCE}
\title{Creates a `TreeViz`` object from `SingleCellExperiment`}
\usage{
createFromSCE(object)
}
\arguments{
\item{object}{`SingleCellExperiment` class containing cluster information at different resolutions}
}
\value{
`TreeViz` Object
}
\description{
Creates a `TreeViz`` object from `SingleCellExperiment`
}
