% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{createFromSeurat}
\alias{createFromSeurat}
\title{Creates a `TreeViz` object from `Seurat`}
\usage{
createFromSeurat(object)
}
\arguments{
\item{object}{`Seurat` class containing cluster information at different resolutions}
}
\value{
`TreeViz` Object
}
\description{
Creates a `TreeViz` object from `Seurat`
}
