% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\name{.generate_hierarchy_tree}
\alias{.generate_hierarchy_tree}
\title{generate hierarchy tree}
\usage{
.generate_hierarchy_tree(hierarchy, feature_order)
}
\arguments{
\item{hierarchy}{hierarchy as a data.table}

\item{feature_order}{order of the tree if different from colnames}
}
\description{
generate hierarchy tree
}
