% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\name{.generate_leaf_of_table}
\alias{.generate_leaf_of_table}
\title{generate leaf of table}
\usage{
.generate_leaf_of_table(
  hierarchy_tree,
  node_ids_table,
  nodes_table,
  feature_order
)
}
\arguments{
\item{hierarchy_tree}{hierarchy as a data.table}

\item{node_ids_table}{node ids}

\item{nodes_table}{nodes table}

\item{feature_order}{order of the tree if different from colnames}
}
\description{
generate leaf of table
}
