% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\name{.generate_node_ids}
\alias{.generate_node_ids}
\title{generate node ids in the tree}
\usage{
.generate_node_ids(hierarchy_tree, feature_order)
}
\arguments{
\item{hierarchy_tree}{hierarchy as a data.table}

\item{feature_order}{order of the tree if different from colnames}
}
\description{
generate node ids in the tree
}
