% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\name{.generate_nodes_table}
\alias{.generate_nodes_table}
\title{generate nodes table tree}
\usage{
.generate_nodes_table(hierarchy_tree, node_ids_table, feature_order)
}
\arguments{
\item{hierarchy_tree}{hierarchy as a data.table}

\item{node_ids_table}{node ids}

\item{feature_order}{order of the tree if different from colnames}
}
\description{
generate nodes table tree
}
