% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\name{.replaceNAFeatures}
\alias{.replaceNAFeatures}
\title{replace if there are NA's in the hierarchy}
\usage{
.replaceNAFeatures(replacing_na_obj_fData, feature_order)
}
\arguments{
\item{replacing_na_obj_fData}{hierarchy data table}

\item{feature_order}{order of the tree if different from colnames}
}
\description{
replace if there are NA's in the hierarchy
}
