% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{generate_walktrap_hierarchy}
\alias{generate_walktrap_hierarchy}
\title{Creates hierarchical clustering on `Single Cell Experiment` object via the walktrap algorithm.
walktrap returns clustering at highest modularity. The modularity value indicates quality of cluster division.
Intermediate cluster assignments are created based on monotonically increasing level of modularity}
\usage{
generate_walktrap_hierarchy(object, nsteps = 7)
}
\arguments{
\item{object}{`Single Cell Experiment` on which `WalkTrap` clustering will be computed}

\item{nsteps}{number of steps to use in `cluster_walktrap`}
}
\value{
dataframe with hierarchy information
}
\description{
Creates hierarchical clustering on `Single Cell Experiment` object via the walktrap algorithm.
walktrap returns clustering at highest modularity. The modularity value indicates quality of cluster division.
Intermediate cluster assignments are created based on monotonically increasing level of modularity
}
