% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\docType{methods}
\name{getNodes,TreeIndex-method}
\alias{getNodes,TreeIndex-method}
\title{Method to get nodes at a tree level}
\usage{
\S4method{getNodes}{TreeIndex}(x, selectedLevel = 3, start = 1,
  end = 1000)
}
\arguments{
\item{x}{a TreeIndex object}

\item{selectedLevel}{tree level to select nodes from}

\item{start, end}{indices to filter nodes by}
}
\description{
Method to get nodes at a tree level
}
