% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{getNodes}
\alias{getNodes}
\title{Generic method to get nodes at a tree level}
\usage{
getNodes(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters}
}
\description{
Generic method to get nodes at a tree level
}
