% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{getNodeStates,TreeIndex-method}
\alias{getNodeStates,TreeIndex-method}
\title{Method to get possible node states
a node state is 0 if removed, 1 if expanded to show children &
2 if counts are aggregated to the node}
\usage{
\S4method{getNodeStates}{TreeIndex}(x)
}
\arguments{
\item{x}{object}
}
\value{
node states
}
\description{
Method to get possible node states
a node state is 0 if removed, 1 if expanded to show children &
2 if counts are aggregated to the node
}
