% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{getNodeStates}
\alias{getNodeStates}
\title{Generic method for possible node states}
\usage{
getNodeStates(x)
}
\arguments{
\item{x}{object}
}
\description{
Generic method for possible node states
}
