% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-class.R
\name{ImportFromSeurat}
\alias{ImportFromSeurat}
\title{Import `Seurat` and `Clustree` into `TreeSummarizedExperiment`}
\usage{
ImportFromSeurat(seurat, clustree, cluster_names = NULL)
}
\arguments{
\item{seurat}{seurat object that contains rowData}

\item{clustree}{clustree object in graph format}

\item{cluster_names}{sluter names from seurat if different from standard}
}
\description{
Import `Seurat` and `Clustree` into `TreeSummarizedExperiment`
}
