% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-class.R
\name{ImportFromSingleCellExperiment}
\alias{ImportFromSingleCellExperiment}
\title{Import `SingleCellExperiment` and `Clustree` into `TreeSummarizedExperiment`}
\usage{
ImportFromSingleCellExperiment(scExp, clustree, cluster_names = NULL)
}
\arguments{
\item{scExp}{SingleCellExperiment object that contains colData}

\item{clustree}{clustree object in graph format}

\item{cluster_names}{sluter names from SingleCellExperiment if different from standard}
}
\description{
Import `SingleCellExperiment` and `Clustree` into `TreeSummarizedExperiment`
}
