% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{PBMC3k_TreeViz}
\alias{PBMC3k_TreeViz}
\title{PBMC 3K dataset for use with vignette}
\format{
A \code{\link{TreeViz}} object
}
\source{
PBMC 3K: \url{https://tcga-data.nci.nih.gov/tcga/}
}
\usage{
data(PBMC3k_TreeViz)
}
\description{
PBMC 3K dataset for use with vignette
}
\keyword{datasets}
