% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeViz-methods.R
\name{plot,TreeViz,ANY-method}
\alias{plot,TreeViz,ANY-method}
\title{plot tree from TreeViz}
\usage{
\S4method{plot}{TreeViz,ANY}(x, y)
}
\arguments{
\item{x}{treeviz object}

\item{y}{none}
}
\value{
Dataframe containing cluster information at different resolutions
}
\description{
plot tree from TreeViz
}
