% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{preprocessAndCreateTreeViz}
\alias{preprocessAndCreateTreeViz}
\title{Creates the `TreeViz` object from cluster hierarchy and count matrix.
Also prunes and collapses the hierarchy to remove redundant levels.}
\usage{
preprocessAndCreateTreeViz(clusters, counts)
}
\arguments{
\item{clusters}{Dataframe containing cluster information at different resolutions}

\item{counts}{matrix Dense or sparse matrix containing the count matrix}
}
\value{
TreeSummarizedExperiment Object that can be visualized with metavizr
}
\description{
Creates the `TreeViz` object from cluster hierarchy and count matrix.
Also prunes and collapses the hierarchy to remove redundant levels.
}
