% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{prune_tree}
\alias{prune_tree}
\title{prune tree to remove non-core edges from clustree hierarchy}
\usage{
prune_tree(graph, cluster_df)
}
\arguments{
\item{graph}{graph from clustree}

\item{cluster_df}{Original cluster hierarchy as dataframe}
}
\value{
Dataframe containing cluster information at different resolutions
}
\description{
prune tree to remove non-core edges from clustree hierarchy
}
