% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeViz-methods.R
\name{register,TreeViz-method}
\alias{register,TreeViz-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{TreeViz}(object, tree = "row", columns = NULL, ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{columns}{Name of columns containing data to register}

\item{...}{Additional arguments passed to object constructors}
}
\value{
An \code{\link{EpivizTreeData-class}} object
}
\description{
Generic method to register data to the epiviz data server
}
