% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{rename_clusters}
\alias{rename_clusters}
\title{Renames cluster names for consistency in visualization}
\usage{
rename_clusters(clusterdata)
}
\arguments{
\item{clusterdata}{Dataframe containing cluster information at different resolutions}
}
\value{
Dataframe containing cluster information at different resolutions
}
\description{
Renames cluster names for consistency in visualization
}
