% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{set_gene_list}
\alias{set_gene_list}
\title{Sets gene list for visualization}
\usage{
set_gene_list(treeviz, genes)
}
\arguments{
\item{treeviz}{TreeViz object}

\item{genes}{list of genes to use}
}
\value{
TreeViz object set with gene list
}
\description{
Sets gene list for visualization
}
