% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startTreeViz.R
\name{setTreevizStandalone}
\alias{setTreevizStandalone}
\title{set treeviz app standalone settings}
\usage{
setTreevizStandalone(
  url = "https://github.com/epiviz/epiviz.git",
  branch = "treeviz",
  local_path = NULL,
  non_interactive = FALSE
)
}
\arguments{
\item{url}{(character) github url to use. defaults to (\url{"https://github.com/epiviz/epiviz.git"}).}

\item{branch}{(character) branch on the github repository. defaults to (master).}

\item{local_path}{(character) if you already have a local instance of treeviz and would like to run standalone use this.}

\item{non_interactive}{(logical) don't download repo, used for testing purposes.}
}
\value{
path to the treeviz app git repository
}
\description{
set treeviz app standalone settings
}
\examples{
\dontrun{
  setTreevizStandalone()
}
}
