% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\docType{methods}
\name{show,TreeIndex-method}
\alias{show,TreeIndex-method}
\title{Show the TreeIndex object}
\usage{
\S4method{show}{TreeIndex}(object)
}
\arguments{
\item{object}{TreeIndex object}
}
\description{
Show the TreeIndex object
}
