% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-methods.R
\docType{methods}
\name{show,TreeSummarizedExperiment-method}
\alias{show,TreeSummarizedExperiment-method}
\title{show object}
\usage{
\S4method{show}{TreeSummarizedExperiment}(object)
}
\arguments{
\item{object}{TreeSummarizedExperiment object}
}
\description{
show object
}
