% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\docType{methods}
\name{splitAt,TreeIndex-method}
\alias{splitAt,TreeIndex-method}
\title{splitAt divides the TreeIndex into groups defined by the level,
node selections and filters(start, end)}
\usage{
\S4method{splitAt}{TreeIndex}(x, selectedLevel = 3,
  selectedNodes = NULL, start = 1, end = 1000, format = "list")
}
\arguments{
\item{x}{TreeIndex object}

\item{selectedLevel}{tree level to select nodes from}

\item{selectedNodes}{used to set states on individual nodes to define a cut on the tree}

\item{start, end}{indices to filter nodes by}

\item{format}{return format can be one of "list" or "TreeIndex"}
}
\description{
splitAt divides the TreeIndex into groups defined by the level,
node selections and filters(start, end)
}
