% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{splitAt}
\alias{splitAt}
\title{Generic method to split the tree}
\usage{
splitAt(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters}
}
\value{
a generic
}
\description{
Generic method to split the tree
}
\examples{
\donttest{ 
library(metagenomeSeq)
data(mouseData)
counts <- MRcounts(mouseData)
hierarchy <- fData(mouseData)
tree <- TreeIndex(hierarchy)
splitAt(tree)
}
}
