% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startTreeViz.R
\name{startTreevizStandalone}
\alias{startTreevizStandalone}
\title{Start treeviz app in standalone (locally) and create \code{\link[TreeViz]{TreevizApp}} object to manage connection.}
\usage{
startTreevizStandalone(
  data = NULL,
  register_function = .register_all_treeviz_things,
  delay = 10L,
  use_viewer_option = FALSE,
  ...
)
}
\arguments{
\item{register_function}{(function) function used to register actions and charts on the treeviz app.}

\item{use_viewer_option}{(function) run application in viewer defined by \code{getOption("viewer")}.
This allows standalone app to run in Rstudio's viewer (FALSE by default)}

\item{...}{additional parameters passed to \code{\link[epivizrStandalone]{startStandalone}}.}
}
\value{
An object of class \code{\link[TreeViz]{TreevizApp}}
}
\description{
Start treeviz app in standalone (locally) and create \code{\link[TreeViz]{TreevizApp}} object to manage connection.
}
\examples{

#' # see package vignette for example usage
app <- startTreevizStandalone(non_interactive=TRUE)
app$stop_app()


}
