% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{[,TreeIndex,ANY,ANY,ANY-method}
\alias{[,TreeIndex,ANY,ANY,ANY-method}
\title{Subset TreeIndex}
\usage{
\S4method{[}{TreeIndex,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{TreeIndex to subset}

\item{i, j}{indices to subset or keep}

\item{...}{other params to dataframe subset method}

\item{drop}{drop the dimensions of the object. defaults to FALSE}
}
\value{
a `TreeIndex` subset object
}
\description{
Subset TreeIndex
}
