% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCluster-class.R
\name{TreeCluster}
\alias{TreeCluster}
\title{create a new Treecluster object}
\usage{
TreeCluster(
  hierarchy = NULL,
  filter = NULL,
  keyword = "cluster",
  ordering = NULL
)
}
\arguments{
\item{hierarchy}{hierarchy as a dataFrame}
}
\description{
create a new Treecluster object
}
