% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-class.R
\docType{class}
\name{TreeIndex-class}
\alias{TreeIndex-class}
\title{TreeIndex class to manage and query hierarchical data}
\description{
TreeIndex class to manage and query hierarchical data
}
