% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-class.R
\name{TreeSE}
\alias{TreeSE}
\title{The TreeSE class.}
\usage{
TreeSE(assays = SimpleList(), rowData = TreeIndex(),
  colData = TreeIndex(), ...)
}
\description{
SummarizedExperiment-like class for datasets that have hierarchies on either rowData or colData.
For microbiome data, rowData is a tree hierarchy
For single cell data, colData is a tree hierarchy
}
