% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-class.R
\docType{class}
\name{TreeSummarizedExperiment-class}
\alias{TreeSummarizedExperiment-class}
\title{TreeSummarizedExperiment class wrapper for SummarizedExperiment objects}
\description{
TreeSummarizedExperiment class wrapper for SummarizedExperiment objects
}
