% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE-class.R
\name{TreeSummarizedExperiment}
\alias{TreeSummarizedExperiment}
\title{The TreeSummarizedExperiment class.}
\usage{
TreeSummarizedExperiment(assays = SimpleList(), rowData = NULL,
  colData = NULL, ...)
}
\arguments{
\item{assays}{simple list of counts}

\item{rowData}{rowData}

\item{colData}{colData}

\item{...}{other parameters for SummarizedExperiment}
}
\description{
SummarizedExperiment-like class for datasets that have hierarchies on either rowData or colData.
For microbiome data, rowData is a tree hierarchy
For single cell data, colData is a tree hierarchy
}
