% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeVizApp-class.R
\docType{class}
\name{TreeVizApp-class}
\alias{TreeVizApp-class}
\alias{TreeVizApp}
\title{Class managing connection to metaviz application.}
\description{
Class managing connection to metaviz application.
}
\section{Methods}{

\describe{
\item{\code{explore(data)}}{explore adds facetZoon, heatmap and TSNE to the app.}
}}

