% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeViz-methods.R
\name{aggregateTree}
\alias{aggregateTree}
\title{Method to aggregate a TreeViz object}
\usage{
aggregateTree(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other params to the function}
}
\value{
a generic
}
\description{
Method to aggregate a TreeViz object
}
\examples{
\dontrun{ 
library(metagenomeSeq)
data(mouseData)
counts <- MRcounts(mouseData)
hierarchy <- fData(mouseData)
tree <- TreeIndex(hierarchy)
mbiome <- TreeViz(SimpleList(counts=counts), rowData=tree)
aggregateTree(mbiome)
}
}
