% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{createFromSCE}
\alias{createFromSCE}
\title{Creates a `TreeViz`` object from `SingleCellExperiment`. Generates
clusters based on Walktrap algorithm if no default is provided}
\usage{
createFromSCE(
  object,
  prefix = "cluster",
  check_colData = FALSE,
  reduced_dim = c("TSNE")
)
}
\arguments{
\item{object}{`SingleCellExperiment` object to be visualized}
}
\value{
`TreeViz` Object
}
\description{
Creates a `TreeViz`` object from `SingleCellExperiment`. Generates
clusters based on Walktrap algorithm if no default is provided
}
