% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{createFromSeurat}
\alias{createFromSeurat}
\title{Creates a `TreeViz` object from `Seurat`}
\usage{
createFromSeurat(object, reduced_dim = c("TSNE"))

createFromSeurat(object, reduced_dim = c("TSNE"))
}
\arguments{
\item{object}{`Seurat` class containing cluster information at different resolutions}

\item{reduced_dim}{Vector of Dimensionality reduction information provided in `Seurat` object to be added in `TreeViz` (if exists)}
}
\value{
`TreeViz` Object

`TreeViz` Object
}
\description{
Creates a `TreeViz` object from `Seurat`

Creates a `TreeViz` object from `Seurat`
}
