% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{getNodes,TreeIndex-method}
\alias{getNodes,TreeIndex-method}
\title{Method to get nodes at a tree level}
\usage{
\S4method{getNodes}{TreeIndex}(x, selectedLevel = NULL)
}
\arguments{
\item{x}{a TreeIndex object}

\item{selectedLevel}{tree level to select nodes from}
}
\description{
Method to get nodes at a tree level
}
\examples{
library(metagenomeSeq)
data(mouseData)
counts <- MRcounts(mouseData)
hierarchy <- fData(mouseData)
tree <- TreeIndex(hierarchy)
getNodes(tree)
}
