% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeIndex-methods.R
\name{splitAt}
\alias{splitAt}
\title{Generic method to split the tree}
\usage{
splitAt(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters}
}
\description{
Generic method to split the tree
}
