% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeVizApp-class.R
\docType{class}
\name{TreeVizApp-class}
\alias{TreeVizApp-class}
\alias{TreeVizApp}
\title{Class managing connection to metaviz application.}
\description{
Class managing connection to metaviz application.
}
\section{Methods}{

\describe{
\item{\code{plotGene(gene = NULL, datasource_name = "SCRNA_1")}}{Plot a bar plot for a gene across cell types
\describe{
\item{gene}{gene to extract expression values}
\item{datasource_name}{object to extract from (automatically selected)}
}}
}}

